Properly use .hpp for all prime headers
