+++ CMakeLists.txt
@@ -6,34 +6,36 @@ project(prime-practice-native C CXX)
 
 include_directories(PrimeAPI/include)
 include_directories(src/include)
+include_directories(src)
 include_directories(extern/fmt-7.1.3/include)
 
 
 set(SOURCE_FILES
+        PrimeAPI/include/PrimeAPI.h
+
         PrimeAPI/include/rstl/string.h
         PrimeAPI/include/rstl/rstl.h
-        PrimeAPI/include/PrimeAPI.h
-        src/include/prime/fnAddr.cpp
-        src/include/prime/CStateManager.h
+        PrimeAPI/include/rstl/vector.h
         src/include/dvd.h
-        src/include/os.h
-        src/include/prime/CMemoryCardSys.hpp
-        src/include/MathyTypes.hpp
-        src/include/prime/CSimplePool.hpp
         src/include/types.h
+        src/include/os.h
+
+        src/include/MathyTypes.hpp
+        src/include/prime/fnAddr.cpp
+        src/include/prime/CStateManager.hpp
+        src/include/prime/CMemoryCardSys.hpp
+        src/include/prime/CSimplePool.hpp
         src/include/prime/CPauseScreen.hpp
         src/include/prime/CFinalInput.hpp
         src/include/prime/CSfxManager.hpp
         src/include/prime/CTextGui.hpp
         src/include/prime/CColor.hpp
         src/include/prime/CGraphics.hpp
-        src/include/TextRenderer.hpp
-        src/include/NewPauseScreen.hpp
         src/include/prime/CGameGlobalObjects.hpp
         src/include/prime/CHealthInfo.hpp
-        src/include/prime/CMemory.h
-        src/include/prime/CPlayerGun.h
-        src/include/prime/CPlayerState.h
+        src/include/prime/CMemory.hpp
+        src/include/prime/CPlayerGun.hpp
+        src/include/prime/CPlayerState.hpp
         src/include/prime/CWorld.hpp
         src/include/prime/CPhysicsActor.hpp
         src/include/prime/CPhysicsState.hpp
@@ -42,8 +44,6 @@ set(SOURCE_FILES
         src/include/prime/CMain.hpp
         src/include/prime/CMainFlow.hpp
         src/include/prime/CArchitectureMessage.hpp
-        src/include/GetField.hpp
-        src/include/GX.hpp
         src/include/prime/CMFGame.hpp
         src/include/prime/CIOWinManager.hpp
         src/include/prime/CGameAllocator.hpp
@@ -54,14 +54,16 @@ set(SOURCE_FILES
         src/include/prime/CScriptTrigger.hpp
         src/include/prime/CScriptDock.hpp
         src/include/prime/CEntity.hpp
-        PrimeAPI/include/rstl/rstl.h
-        PrimeAPI/include/rstl/vector.h
         src/include/prime/CFontEndUI.hpp
         src/include/prime/CArchitectureQueue.hpp
         src/include/prime/CScriptRelay.hpp
         src/include/prime/CScriptDoor.hpp
         src/include/prime/CScriptCameraHint.hpp
         src/include/STriggerRenderConfig.hpp
+        src/include/TextRenderer.hpp
+        src/include/NewPauseScreen.hpp
+        src/include/GetField.hpp
+        src/include/GX.hpp
         src/new.cpp
         src/NewPauseScreen.cpp
         src/TextRenderer.cpp
