+++ src/NewPauseScreen.cpp
@@ -6,11 +6,13 @@
 #include "os.h"
 #include "NewPauseScreen.hpp"
 #include "TextRenderer.hpp"
+#include "prime/CGameState.hpp"
+#include "prime/CWorldState.hpp"
 #include "prime/CScriptRelay.hpp"
 #include "prime/CScriptDock.hpp"
 #include "prime/CGameGlobalObjects.hpp"
 #include "prime/CPlayer.hpp"
-#include "prime/CPlayerState.h"
+#include "prime/CPlayerState.hpp"
 #include "prime/CWorld.hpp"
 #include "prime/CMain.hpp"
 #include "prime/CSfxManager.hpp"
