+++ src/include/NewPauseScreen.hpp
@@ -4,7 +4,7 @@
 #include <prime/CScriptTrigger.hpp>
 #include <prime/CFinalInput.hpp>
 #include <UI/Menu.h>
-#include "prime/CStateManager.h"
+#include "prime/CStateManager.hpp"
 #include "STriggerRenderConfig.hpp"
 
 void warp(uint32_t world, uint32_t area);
