+++ src/include/prime/CGameState.hpp
@@ -3,16 +3,17 @@
 
 class CWorldState;
 
-class CGameState
-{
+class CGameState {
 public:
-	void SetCurrentWorldId(uint assetId);
-	CWorldState* StateForWorld(uint assetId);
-	CWorldState* CurrentWorldState();
-	
+    void SetCurrentWorldId(CAssetId id);
+    CWorldState &StateForWorld(unsigned int world);
+    CWorldState &CurrentWorldState();
+
+
     inline CAssetId MLVL() { return *GetField<u32>(this, 0x84); };
     inline double PlayTime() { return *GetField<double>(this, 0xa0); };
 };
+
 extern CGameState* g_GameState;
 
 #endif // CGAMESTATE_HPP
